backend/app/services/crawler/workday.py
import logging
from typing import List, Dict, Any
from bs4 import BeautifulSoup
from playwright.async_api import async_playwright

logger = logging.getLogger(__name__)

async def crawl_workday(base_url: str) -> List[Dict[str, Any]]:
    """Crawl Workday job board (requires Playwright for JS rendering)"""
    jobs = []
    try:
        async with async_playwright() as p:
            browser = await p.chromium.launch(headless=True)
            page = await browser.new_page()
            await page.goto(base_url, wait_until="networkidle", timeout=30000)
            
            content = await page.content()
            soup = BeautifulSoup(content, 'html.parser')
            
            # Placeholder: Adjust selectors based on actual Workday structure
            job_items = soup.find_all('div', class_='job-item')
            
            for item in job_items:
                try:
                    title_elem = item.find('a', class_='job-title')
                    location_elem = item.find('span', class_='job-location')
                    
                    job_data = {
                        "external_id": item.get('data-job-id'),
                        "title": title_elem.text.strip() if title_elem else "N/A",
                        "company": "Workday",
                        "location": location_elem.text.strip() if location_elem else "Remote",
                        "description": item.get_text(),
                        "job_type": "full_time",
                        "url": title_elem['href'] if title_elem and title_elem.has_attr('href') else base_url,
                        "posting_date": None
                    }
                    jobs.append(job_data)
                except Exception as e:
                    logger.warning(f"Error parsing Workday job: {e}")
                    continue
            
            await browser.close()
    
    except Exception as e:
        logger.error(f"Workday crawl error: {e}")
    
    return jobs


