backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic>=2.9.0
pydantic-settings>=2.1.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography>=41.0.0
alembic==1.13.0

# Crawling
playwright==1.40.0
httpx==0.25.2
beautifulsoup4==4.12.3
lxml>=5.3.0

# LLM & RAG
langchain==0.1.12
numpy>=1.24,<2
langchain-openai==0.0.5
openai>=1.10.0,<2.0.0
sentence-transformers==3.1.1
reportlab==4.2.5
faiss-cpu==1.8.0
# chromadb==0.4.17   # removed: incompatible with pydantic v2
# crewai>=1.0.0      # removed: not used currently
# crewai-tools>=0.1.0

# Scheduler
apscheduler==3.10.4

# Utils
pyyaml==6.0.1
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
python-docx==1.1.0
PyPDF2>=3.0.0
redis==5.0.1
openpyxl==3.1.5

# Security
pyjwt>=2.9.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# Logging
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0

